third_party/blink/web_tests/platform/generic/fast/events/touch/touch-browser-zoom-scales-radius-expected.txt

This tests makes sure the touch event's radius property is scaled to account for browser zoom.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

===Initial Zoom===
PASS radiusX is 4
PASS radiusY is 10
===Zoom 2X===
PASS radiusX is 2
PASS radiusY is 5
===Zoom 0.5X===
PASS radiusX is 8
PASS radiusY is 20
PASS successfullyParsed is true

TEST COMPLETE
